place/requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
Werkzeug==0.12.2
